Refactoring OpenShiftOAuth
