+++ controllers/che/checluster_controller_test.go
@@ -18,7 +18,6 @@ import (
 	"strconv"
 
 	mocks "github.com/eclipse-che/che-operator/mocks"
-	"github.com/stretchr/testify/assert"
 
 	"reflect"
 	"time"
@@ -70,45 +69,7 @@ import (
 )
 
 var (
-	namespace        = "eclipse-che"
-	nonEmptyUserList = &userv1.UserList{
-		Items: []userv1.User{
-			{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: "user1",
-				},
-			},
-			{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: "user2",
-				},
-			},
-		},
-	}
-	oAuthWithNoIdentityProviders = &configv1.OAuth{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: "cluster",
-			// Namespace: namespace,
-		},
-	}
-	oAuthWithIdentityProvider = &configv1.OAuth{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: "cluster",
-			// Namespace: namespace,
-		},
-		Spec: configv1.OAuthSpec{
-			IdentityProviders: []configv1.IdentityProvider{
-				{
-					Name: "htpasswd",
-				},
-			},
-		},
-	}
-	proxy = &configv1.Proxy{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: "cluster",
-		},
-	}
+	namespace = "eclipse-che"
 )
 
 func TestNativeUserModeEnabled(t *testing.T) {
@@ -223,179 +184,6 @@ func TestNativeUserModeEnabled(t *testing.T) {
 				t.Errorf("Expected nativeUserMode '%+v', but found '%+v' for input '%+v'",
 					expectedValue, actualValue, testCase)
 			}
-		})
-	}
-}
-
-func TestCaseAutoDetectOAuth(t *testing.T) {
-	type testCase struct {
-		name                             string
-		initObjects                      []runtime.Object
-		isOpenshift4                     bool
-		initialOAuthValue                *bool
-		oAuthExpected                    *bool
-		initialOpenShiftOAuthUserEnabled *bool
-	}
-
-	testCases := []testCase{
-		{
-			name: "che-operator should auto enable oAuth when Che CR with oAuth nil value on the Openshift 3 with users > 0",
-			initObjects: []runtime.Object{
-				nonEmptyUserList,
-				&oauthv1.OAuthClient{},
-			},
-			isOpenshift4:      false,
-			initialOAuthValue: nil,
-			oAuthExpected:     pointer.BoolPtr(true),
-		},
-		{
-			name: "che-operator should auto disable oAuth when Che CR with nil oAuth on the Openshift 3 with no users",
-			initObjects: []runtime.Object{
-				&userv1.UserList{},
-				&oauthv1.OAuthClient{},
-			},
-			isOpenshift4:      false,
-			initialOAuthValue: pointer.BoolPtr(false),
-			oAuthExpected:     pointer.BoolPtr(false),
-		},
-		{
-			name: "che-operator should respect oAuth = true even if there no users on the Openshift 3",
-			initObjects: []runtime.Object{
-				&userv1.UserList{},
-				&oauthv1.OAuthClient{},
-			},
-			isOpenshift4:      false,
-			initialOAuthValue: pointer.BoolPtr(true),
-			oAuthExpected:     pointer.BoolPtr(true),
-		},
-		{
-			name: "che-operator should respect oAuth = true even if there are some users on the Openshift 3",
-			initObjects: []runtime.Object{
-				nonEmptyUserList,
-				&oauthv1.OAuthClient{},
-			},
-			isOpenshift4:      false,
-			initialOAuthValue: pointer.BoolPtr(true),
-			oAuthExpected:     pointer.BoolPtr(true),
-		},
-		{
-			name: "che-operator should respect oAuth = false even if there are some users on the Openshift 3",
-			initObjects: []runtime.Object{
-				nonEmptyUserList,
-				&oauthv1.OAuthClient{},
-			},
-			isOpenshift4:      false,
-			initialOAuthValue: pointer.BoolPtr(false),
-			oAuthExpected:     pointer.BoolPtr(false),
-		},
-		{
-			name: "che-operator should respect oAuth = false even if no users on the Openshift 3",
-			initObjects: []runtime.Object{
-				&userv1.UserList{},
-				&oauthv1.OAuthClient{},
-			},
-			isOpenshift4:      false,
-			initialOAuthValue: pointer.BoolPtr(false),
-			oAuthExpected:     pointer.BoolPtr(false),
-		},
-		{
-			name: "che-operator should auto enable oAuth when Che CR with nil value on the Openshift 4 with identity providers",
-			initObjects: []runtime.Object{
-				oAuthWithIdentityProvider,
-				proxy,
-			},
-			isOpenshift4:      true,
-			initialOAuthValue: nil,
-			oAuthExpected:     pointer.BoolPtr(true),
-		},
-		{
-			name: "che-operator should respect oAuth = true even if there no indentity providers on the Openshift 4",
-			initObjects: []runtime.Object{
-				oAuthWithNoIdentityProviders,
-				proxy,
-			},
-			isOpenshift4:                     true,
-			initialOAuthValue:                pointer.BoolPtr(true),
-			oAuthExpected:                    pointer.BoolPtr(true),
-			initialOpenShiftOAuthUserEnabled: pointer.BoolPtr(true),
-		},
-		{
-			name: "che-operator should respect oAuth = true even if there are some users on the Openshift 4",
-			initObjects: []runtime.Object{
-				oAuthWithIdentityProvider,
-				proxy,
-			},
-			isOpenshift4:                     false,
-			initialOAuthValue:                pointer.BoolPtr(true),
-			oAuthExpected:                    pointer.BoolPtr(true),
-			initialOpenShiftOAuthUserEnabled: pointer.BoolPtr(true),
-		},
-		{
-			name: "che-operator should respect oAuth = false even if there no indentity providers on the Openshift 4",
-			initObjects: []runtime.Object{
-				oAuthWithNoIdentityProviders,
-				proxy,
-			},
-			isOpenshift4:      false,
-			initialOAuthValue: pointer.BoolPtr(false),
-			oAuthExpected:     pointer.BoolPtr(false),
-		},
-		{
-			name: "che-operator should respect oAuth = false even if there are some users on the Openshift 4",
-			initObjects: []runtime.Object{
-				oAuthWithIdentityProvider,
-				proxy,
-			},
-			isOpenshift4:      false,
-			initialOAuthValue: pointer.BoolPtr(false),
-			oAuthExpected:     pointer.BoolPtr(false),
-		},
-		{
-			name:                             "che-operator should auto disable oAuth on error retieve identity providers",
-			initObjects:                      []runtime.Object{},
-			isOpenshift4:                     false,
-			initialOAuthValue:                nil,
-			initialOpenShiftOAuthUserEnabled: pointer.BoolPtr(true),
-			oAuthExpected:                    pointer.BoolPtr(false),
-		},
-	}
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			logf.SetLogger(zap.New(zap.WriteTo(os.Stdout), zap.UseDevMode(true)))
-
-			initCR := InitCheWithSimpleCR().DeepCopy()
-			initCR.Spec.Auth.OpenShiftoAuth = testCase.initialOAuthValue
-			initCR.Spec.Auth.InitialOpenShiftOAuthUser = testCase.initialOpenShiftOAuthUserEnabled
-
-			deployContext := deploy.GetTestDeployContext(initCR, testCase.initObjects)
-
-			r := NewReconciler(
-				deployContext.ClusterAPI.Client,
-				deployContext.ClusterAPI.NonCachedClient,
-				deployContext.ClusterAPI.DiscoveryClient,
-				deployContext.ClusterAPI.Client.Scheme(),
-				"")
-			r.tests = true
-
-			req := reconcile.Request{
-				NamespacedName: types.NamespacedName{
-					Name:      os.Getenv("CHE_FLAVOR"),
-					Namespace: namespace,
-				},
-			}
-
-			util.IsOpenShift = true
-			util.IsOpenShift4 = testCase.isOpenshift4
-
-			_, err := r.Reconcile(context.TODO(), req)
-			assert.Nil(t, err)
-
-			cheCR := &orgv1.CheCluster{}
-			err = r.client.Get(context.TODO(), types.NamespacedName{Name: os.Getenv("CHE_FLAVOR"), Namespace: namespace}, cheCR)
-			assert.Nil(t, err)
-
-			assert.NotNil(t, cheCR.Spec.Auth.OpenShiftoAuth)
-			assert.Equal(t, *testCase.oAuthExpected, *cheCR.Spec.Auth.OpenShiftoAuth)
 		})
 	}
 }
