+++ controllers/che/create.go
@@ -12,6 +12,8 @@
 package che
 
 import (
+	"strconv"
+
 	"github.com/eclipse-che/che-operator/pkg/deploy"
 	"github.com/eclipse-che/che-operator/pkg/util"
 	"github.com/sirupsen/logrus"
@@ -137,7 +139,7 @@ func (r *CheClusterReconciler) GenerateAndSaveFields(deployContext *deploy.Deplo
 		}
 	}
 
-	if !util.IsNativeUserModeEnabled(deployContext.CheCluster) {
+	if !util.IsOpenShift || !deployContext.CheCluster.IsNativeUserModeEnabled() {
 		keycloakRealm := util.GetValue(deployContext.CheCluster.Spec.Auth.IdentityProviderRealm, cheFlavor)
 		if len(deployContext.CheCluster.Spec.Auth.IdentityProviderRealm) < 1 {
 			deployContext.CheCluster.Spec.Auth.IdentityProviderRealm = keycloakRealm
@@ -242,6 +244,14 @@ func (r *CheClusterReconciler) GenerateAndSaveFields(deployContext *deploy.Deplo
 		strategy := util.GetServerExposureStrategy(deployContext.CheCluster)
 		deployContext.CheCluster.Spec.Server.ServerExposureStrategy = strategy
 		if err := deploy.UpdateCheCRSpec(deployContext, "serverExposureStrategy", strategy); err != nil {
+			return err
+		}
+	}
+
+	if util.IsOpenShift && deployContext.CheCluster.Spec.DevWorkspace.Enable && deployContext.CheCluster.Spec.Auth.NativeUserMode == nil {
+		newNativeUserModeValue := util.NewBoolPointer(true)
+		deployContext.CheCluster.Spec.Auth.NativeUserMode = newNativeUserModeValue
+		if err := deploy.UpdateCheCRSpec(deployContext, "nativeUserMode", strconv.FormatBool(*newNativeUserModeValue)); err != nil {
 			return err
 		}
 	}
